Initial load of several YARP modules for IDSIA
The idsiayarpmodules directory tree contains source code, documentation, build scripts and
other supporting files for a number of extension modules for YARP, developed for use in
IDSIA. It is supposed to be a resting place for additional YARP modules.

It initially contains the following modules / functionality:
- A driver for Firewire cameras
- A viewer for YARP frame-grappers
- A driver for Katana 400 robots
- Several filters that can be used to observe and intercept
  data and/or commands flowing in/out of YARP ports. These
  filters are intended to be used as hook-in points for
  e.g. collision avoidance control

The initial content of this directory tree has been developed in the course of my
master thesis at IDSIA's robotics lab.